% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{get.registration.numbers}
\alias{get.registration.numbers}
\title{Extract all registration numbers}
\usage{
get.registration.numbers(folder, ...)
}
\arguments{
\item{folder}{Folder where all ibutton files are kept}
}
\description{
reads all the ibutton files in one folder,
read only the top of each file and extract from that the
registration numbers
}

