% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ibutton_functions.R
\name{get_preamble_length}
\alias{get_preamble_length}
\title{Get the length of the 'preamble' of all ibuttons}
\usage{
get_preamble_length(allfiles)
}
\arguments{
\item{allfiles}{a list of all file names of ibuttons}
}
\value{
an integer giving the number of lines before the header row of the actual data
}
\description{
ibuttons produce csvs with several lines of text at the beginning,
which describe how the ibutton was configured. This function
determines the length of this text or "preamble".
}

