% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ibutton_functions.R
\name{get_registration_numbers}
\alias{get_registration_numbers}
\title{Extract all registration numbers}
\usage{
get_registration_numbers(folder, ...)
}
\arguments{
\item{folder}{Folder where all ibutton files are kept}
}
\description{
reads all the ibutton files in one folder,
read only the top of each file and extract from that the
registration numbers
}

