% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ibutton_functions.R
\name{header_names}
\alias{header_names}
\title{Create more accurate header names for the dataset}
\usage{
header_names(allfiles)
}
\arguments{
\item{allfiles}{a list of all file names of ibuttons}
}
\value{
an integer giving the number of fields in row below the header
}
\description{
Occasionally ibuttons produce data with other, unlabelled numbers in a column
to the right of the actual data. This function counts the fields in the row below the data
and then supplies a nicer header
}

