% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ibuttons.to.data.frame}
\alias{ibuttons.to.data.frame}
\title{Combines given list of ibuttons into a data.frame}
\usage{
ibuttons.to.data.frame(ibut.list)
}
\arguments{
\item{ibut.list}{List of ibutton data}
}
\description{
This function takes a list of ibuttons and \code{rbinds} them.
Also performs date conversion
}

