% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{id.broken}
\alias{id.broken}
\title{Check for broken ibuttons}
\usage{
id.broken(ibutt.list)
}
\arguments{
\item{ibutt.list}{List of ibuttons to combine}
}
\value{
names of suspect ibuttons
}
\description{
ibuttons sometimes break, ie. stop recording. This function tries to
identify those
}

