% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ibutton_functions.R
\name{id_broken}
\alias{id_broken}
\title{Check for broken ibuttons}
\usage{
id_broken(ibutt.list)
}
\arguments{
\item{ibutt.list}{List of ibuttons to combine}
}
\value{
names of suspect ibuttons
}
\description{
ibuttons sometimes break, ie. stop recording. This function tries to
identify those
}

