% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{preamble.extract.registration.number}
\alias{preamble.extract.registration.number}
\title{Extracts the registration number for an ibutton}
\usage{
preamble.extract.registration.number(ibutton.preamble)
}
\arguments{
\item{ibutton.preamble}{An ibutton preamble}
}
\value{
Registration number (character)
}
\description{
This function extracts the  from the "preamble", ie the 13 lines
of information at the start of an ibutton file
}

