% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ibutton_functions.R
\name{preamble_extract_registration_number}
\alias{preamble_extract_registration_number}
\title{Extracts the registration number for an ibutton}
\usage{
preamble_extract_registration_number(ibutton.preamble)
}
\arguments{
\item{ibutton.preamble}{An ibutton preamble}
}
\value{
Registration number (character)
}
\description{
This function extracts the registration number from the "preamble", ie the lines
of information at the start of an ibutton file
}

