% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{read.csv.with.name}
\alias{read.csv.with.name}
\title{Add filename to data}
\usage{
read.csv.with.name(f, ...)
}
\arguments{
\item{f}{Filename of ibutton data}
}
\value{
a \code{data.frame} with a column for original file name
}
\description{
Add filename to data
}

