% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{read.ibutton.folder}
\alias{read.ibutton.folder}
\title{Read all files in a folder}
\usage{
read.ibutton.folder(folder, ...)
}
\arguments{
\item{folder}{Folder containing ibutton data only}
}
\value{
Named list of ibutton data. Names are taken from base names.
}
\description{
loads in all the data held in a single folder.
}
\details{
This function assumes that you have one folder containing only ibutton data.
}

