% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{ibuttonr}
\alias{ibuttonr}
\alias{ibuttonr-package}
\title{Read data from ibuttons.}
\description{
ibuttonr contains some convenience functions for loading
data from ibuttons
}
\details{
The functions read, clean and check the quality of your data
}

