% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ibutton_functions.R
\name{read_ibutton_folder}
\alias{read_ibutton_folder}
\title{Read all files in a folder}
\usage{
read_ibutton_folder(folder, ...)
}
\arguments{
\item{folder}{Folder containing ibutton data only}
}
\value{
Named list of ibutton data. Names are taken from base names.
}
\description{
loads in all the data held in a single folder.
}
\details{
This function assumes that you have one folder containing only ibutton data.
}

