% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_named_ibutton}
\alias{read_named_ibutton}
\title{Add filename to data}
\usage{
read_named_ibutton(f, ...)
}
\arguments{
\item{f}{Filename of ibutton data}
}
\value{
a \code{data.frame} with a column for original file name
}
\description{
Add filename to data
}

